api: phase 4b — hoist shared step-option structs into pipeline::common
Consolidate the byte-identical per-problem step-option structs into one
canonical definition each in a new `vision-calibration-pipeline::common`
module. The intrinsics init/optimize and hand-eye init/optimize options
were duplicated across 5 and 2 problem modules respectively; each problem
now re-exports the shared type so existing paths still resolve, and the
contract can no longer silently diverge per problem.
